Gen encryption key on login + show file metadata
Firstly, we're now generating an encryption key whenever you login to
the server, and storing it in localStorage. This is useful for whenver
we want to decrypt stuff in typescript (all the time).

We're also actually showing the decrypted file names in `/files`! We do
this by downloading the file metadata from Phoenix, then decrypting it
in typescript.

Both of these use the Phoenix `push_event` functionality to easily
update files / know when to create the encryption key. Elixir Phoenix
ftw! We still need to make sure to never pass the raw password to the
server though (which we're doing ATM) as this literally breaks
everything we have set up for encryption.