Allow uploading file contents
We now have an almost complete ugly client! We do some wasm magic to
encrypt and compress file chunks, and send them up to the server.
Haven't even tried testing downloading yet, but that's something for
another time. Also we're base64 encoding any binary data we upload >:(.
This is absolutely not necessary and something we need to *not* do, but
i want to get a working prototype out soon.