connect directly from browser to efs
i've been thinking about it, and i don't see why we need to proxy from
big_central to efs. it's less efficient, and introduces another point of
failure. it's way better to keep the entire efs client in typescript,
and then connect directly to it.

the way we're doing that is using webtransport. i've been meaning to
switch the server over to QUIC from TCP anyway, so this works out. the
main downside is that webtransport is a PAIN to host locally, but i'll
add guides soon enough. anyway most of the code changes involve adding
the bfsp and efs clients, using webtransport to connect to the server,
and adding a bit more functionality to typescript and wasm.