Add the ability to login + hash passwords client side
Firstly, we added the ability to login to your account! Very advanced
stuff, I know.

We're also hashing passwords client side. The reason we do this is that
we generate the encryption key based on the user's password, and as such
we don't want to let the server see that password in plain text ever.