Added /files/view_file
This is basically the file sharing / file view page. We generate a new
token that's restricted to only being allowed to access a single file.
We also make sure to put the file's encryption key, so users can
download it. Finally, all of this info is stored in the URI fragment
identifier, so that the server isn't able to access any of it.

I alos did some refactoring to make code nicer and easier to read, while
subsequently making it grosser.

Finally, I added a little loading indicator, since in testing users
click off of a slower loading page.