app authentication now sends the master key
having the user type in their password to gen an encryption key is
dogshit UX. the app now sends up a b64 encoded rsa public key alongside
their dl token. the typescript client then uses that rsa public key to
encrypt the master encryption key, and store that on the server
temporarily. the server then offers up the user's token, as well as the
user's encrypted master key. it's then the job of the client to decrypt
that master key to begin using the app